% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{add_decomposition}
\alias{add_decomposition}
\title{Add STL Decomposition Components}
\usage{
add_decomposition(tsbl, long = FALSE)
}
\arguments{
\item{tsbl}{tsibble}

\item{long}{return a long tsibble with only one measured variable}
}
\value{
tsibble with columns for each component of the decomposition, including the residual
}
\description{
Add STL Decomposition Components
}
