% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{clean_tsbl}
\alias{clean_tsbl}
\title{Ungroup and fix frequency of tsibble}
\usage{
clean_tsbl(tsbl)
}
\arguments{
\item{tsbl}{tsibble object}
}
\value{
ungrouped tsibble object of correct frequency
}
\description{
Ungroup and fix frequency of tsibble
}
