% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{correct_frequency}
\alias{correct_frequency}
\title{Correct the frequency of a tsibble}
\usage{
correct_frequency(tsbl)
}
\arguments{
\item{tsbl}{tsibble}
}
\value{
tsibble
}
\description{
Correct the frequency of a tsibble
}
