% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{impute_values}
\alias{impute_values}
\title{Add Imputed Series}
\usage{
impute_values(
  tsbl,
  method = c("linear", "spline", "stine", "ma", "StructTS", "ARIMA", "locf", "nocb",
    "mean", "median", "mode", "random", "replace", "seasonal_interpolation",
    "seasonal_locf", "seasonal_mean", "seasonal_random", "seasonal_kalman", "seasonal_ma"),
  long = FALSE,
  value = NULL,
  Kalman_smooth = FALSE,
  seasonal_splitting = FALSE
)
}
\arguments{
\item{tsbl}{tsibble}

\item{method}{method of imputation}

\item{long}{return a long tsibble with only one measured variable}

\item{value}{if method is "replace" then the value for replacement. if "ma" then the value of k.}

\item{Kalman_smooth}{use Kalman smoothing for StructTS or ARIMA}

\item{seasonal_splitting}{}
}
\value{
tsibble with a column for imputed values
}
\description{
Add Imputed Series
}
