% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Add Series without Outliers}
\usage{
remove_outliers(tsbl, long = FALSE)
}
\arguments{
\item{tsbl}{tsibble}

\item{long}{return a long tsibble with only one measured variable}
}
\value{
tsibble with a column for cleaned values
}
\description{
Add Series without Outliers
}
