% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{stash}
\alias{stash}
\title{Store an object to the global env mid-pipeline}
\usage{
stash(.data, name)
}
\arguments{
\item{.data}{an object to be assigned to name}

\item{name}{variable name}
}
\value{
.data argument, to allow continued piping, with the side effect of assigning the .data object to an object using the name argument
}
\description{
Store an object to the global env mid-pipeline
}
